+++ src/cli/programutilities.h
@@ -8,4 +8,6 @@ namespace Ripes {
 QString loadFlatBinaryFile(Program &program, const QString &filepath,
                            unsigned long entryPoint, unsigned long loadAt);
 
+bool loadElfFile(Program &program, QFile &file);
+
 } // namespace Ripes
