+++ src/syscall/file.h
@@ -87,7 +87,7 @@ public:
                     {{0, "number of read bytes or -1 if an error occurred"}}) {}
   void execute() {
     const int fd = BaseSyscall::getArg(BaseSyscall::REG_FILE, 0);
-    int byteAddress = BaseSyscall::getArg(
+    long byteAddress = BaseSyscall::getArg(
         BaseSyscall::REG_FILE, 1); // destination of characters read from file
     const int length = BaseSyscall::getArg(BaseSyscall::REG_FILE, 2);
     QByteArray buffer;
@@ -120,7 +120,7 @@ public:
                      {2, "number of bytes to write"}},
                     {{0, "the number of bytes written"}}) {}
   void execute() {
-    const int byteAddress = BaseSyscall::getArg(
+    const long byteAddress = BaseSyscall::getArg(
         BaseSyscall::REG_FILE, 1); // source of characters to write to file
     const int reqLength =
         BaseSyscall::getArg(BaseSyscall::REG_FILE, 2); // user-requested length
@@ -156,7 +156,7 @@ public:
              {1, "the length of the buffer"}},
             {{0, "-1 if the path is longer than the buffer"}}) {}
   void execute() {
-    const int byteAddress = BaseSyscall::getArg(
+    const long byteAddress = BaseSyscall::getArg(
         BaseSyscall::REG_FILE, 0); // destination of characters read from file
     int index = 0;
     const int bufferSize = BaseSyscall::getArg(BaseSyscall::REG_FILE, 1);
